query_runner/requirements.txt
kafka-python==2.0.2
kafka-python==2.0.2
requests